requirements.txt

Django==5.2.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1